+++ src/main/java/nextstep/subway/section/domain/Section.java
@@ -1,14 +1,13 @@
 package nextstep.subway.section.domain;
 
-import static nextstep.subway.section.domain.exception.StationExceptionMessage.CANNOT_EQUALS_UP_STATION_WITH_DOWN_STATION;
-import static nextstep.subway.section.domain.exception.StationExceptionMessage.DISTANCE_IS_NOT_NULL;
-import static nextstep.subway.section.domain.exception.StationExceptionMessage.DOWN_STATION_IS_NOT_NULL;
-import static nextstep.subway.section.domain.exception.StationExceptionMessage.UP_STATION_IS_NOT_NULL;
+import static nextstep.subway.station.domain.exception.StationExceptionMessage.CANNOT_EQUALS_UP_STATION_WITH_DOWN_STATION;
+import static nextstep.subway.station.domain.exception.StationExceptionMessage.DISTANCE_IS_NOT_NULL;
+import static nextstep.subway.station.domain.exception.StationExceptionMessage.DOWN_STATION_IS_NOT_NULL;
+import static nextstep.subway.station.domain.exception.StationExceptionMessage.UP_STATION_IS_NOT_NULL;
 
 import java.util.Objects;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.ForeignKey;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
